+++ packages/select/src/components/base-select/components/list-mobile/list-modal-mobile.tsx
@@ -2,8 +2,9 @@ import React, { FC, ReactNode, RefObject } from 'react';
 import mergeRefs from 'react-merge-refs';
 import cn from 'classnames';
 
-import { OptionShape } from '@alfalab/core-components-select/typings';
 import { getDataTestId } from '@alfalab/core-components-shared';
+
+import { OptionShape } from '../../../../typings';
 
 import { ListModalMobileRestProps } from './types/types';